requirements.txt
pytz
pycurl
requests
jinja2
facebook-sdk
